break out browser specific parts, use hyperquest in node
